+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 unreleased
 - CO-2876 implement model count
 -----
+- CO-3165 add IdentityType for deleted product image
+
 
 5.3.0
 -----